インストーラーを見直し

+++ install.sh
@@ -27,7 +27,9 @@ curl --fail --location --progress-bar --output /tmp/ezcron.tar.gz ${ezjob_url}
 cd /tmp
 sudo tar zxf ezcron.tar.gz -C /tmp
 sudo cp ezcron /usr/local/bin/
-sudo cp ezcron.toml /etc/
+if [ ! -f /etc/ezcron.toml ]; then
+    sudo cp ezcron.toml /etc/
+fi
 sudo rm /tmp/ezcron
 sudo rm /tmp/ezcron.toml
 sudo rm /tmp/ezcron.tar.gz
